Clean up loan support API
