+++ src/core/ddsc/include/dds/ddsc/dds_data_allocator.h
@@ -112,31 +112,6 @@ DDS_EXPORT void *dds_data_allocator_alloc (dds_data_allocator_t *data_allocator,
  */
 DDS_EXPORT dds_return_t dds_data_allocator_free (dds_data_allocator_t *data_allocator, void *ptr);
 
-/** @brief Check if a Loan is available to reader/writer
- * The loan is available if the shared memory is enabled and all the constraints to enable
- * shared memory are met and the type is fixed
- *
- * @param[in] entity the handle of the entity
- *
- * @returns loan available or not
- */
-DDS_EXPORT bool dds_is_loan_available (const dds_entity_t entity);
-
-/** @brief Check if a shared memory is available to reader/writer.
- *
- * @note shared memory available and loan available have different meaning
- * 1)We may not be able to use the loan API due to type constraints (not fixed)
- * 2) We may also allow using the loan API but without shared memory
- * but in this case not zero-copy.
- *
- * @param[in] entity the handle of the entity
- *
- * @returns true if shared memory is available, false otherwise
- */
-DDS_EXPORT bool dds_is_shared_memory_available(const dds_entity_t entity);
-
-DDS_DEPRECATED_EXPORT bool is_loan_available (const dds_entity_t entity);
-
 #if defined (__cplusplus)
 }
 #endif
