+++ src/core/ddsc/include/dds/ddsc/dds_public_impl.h
@@ -32,6 +32,8 @@
 extern "C" {
 #endif
 
+typedef int32_t dds_entity_t;
+
 typedef struct dds_sequence
 {
   uint32_t _maximum;
