+++ src/core/ddsc/src/dds_data_allocator.c
@@ -16,6 +16,8 @@
 #include "dds__data_allocator.h"
 #include "dds__entity.h"
 
+#include "dds/ddsc/dds_loan.h"
+
 dds_return_t dds_data_allocator_init_heap (dds_data_allocator_t *data_allocator)
 {
   // Use special entity handle to allocate on heap
@@ -108,10 +110,9 @@ void *dds_data_allocator_alloc (dds_data_allocator_t *data_allocator, size_t siz
       if (size > UINT32_MAX)
         return NULL;
       else {
-        // MAKI: We get problems with bookkeeping of loans etc., for which we
-        // need the dds_writer. We also need to lock the access to the iceoryx
-        // publisher if this is to be used concurrently. Currently we use the
-        // writer lock for this.
+        // TODO: we need to lock the iceoryx publisher with an extra lock
+        // It is not thread-safe on its own. The allocator must own the
+        // corresponding lock.
         return shm_create_chunk(d->ref.pub, (uint32_t)size);
       }
     default:
@@ -158,78 +159,4 @@ dds_return_t dds_data_allocator_free (dds_data_allocator_t *data_allocator, void
   ddsrt_free (ptr);
 #endif
   return ret;
-}
-
-// MAKI: return memory property struture instead? (dynamic, shared memory, other
-// ...), rename into memory properties in this case
-bool dds_is_shared_memory_available(const dds_entity_t entity) {
-  bool ret = false;
-#ifdef DDS_HAS_SHM
-  dds_entity *e;
-
-  if (DDS_RETCODE_OK != dds_entity_pin(entity, &e)) {
-    return ret;
-  }
-
-  switch (dds_entity_kind(e)) {
-  case DDS_KIND_READER: {
-    struct dds_reader const *const rd = (struct dds_reader *)e;
-    // only if SHM is enabled correctly (i.e. iox subscriber is initialized) and
-    // the type is fixed
-    ret = (rd->m_iox_sub != NULL);
-    break;
-  }
-  case DDS_KIND_WRITER: {
-    struct dds_writer const *const wr = (struct dds_writer *)e;
-    // only if SHM is enabled correctly (i.e. iox publisher is initialized) and
-    // the type is fixed
-    ret = (wr->m_iox_pub != NULL);
-    break;
-  }
-  default:
-    break;
-  }
-
-  dds_entity_unpin(e);
-#endif
-  (void)entity;
-  return ret;
-}
-
-bool dds_is_loan_available(const dds_entity_t entity)
-{
-  bool ret = false;
-#ifdef DDS_HAS_SHM
-  dds_entity * e;
-
-  if (DDS_RETCODE_OK != dds_entity_pin(entity, &e)) {
-    return ret;
-  }
-
-  switch (dds_entity_kind(e)) {
-    case DDS_KIND_READER: {
-      struct dds_reader const *const rd = (struct dds_reader *)e;
-      // only if SHM is enabled correctly (i.e. iox subscriber is initialized) and the type is fixed
-      ret = (rd->m_iox_sub != NULL) && (rd->m_topic->m_stype->fixed_size);
-      break;
-    }
-    case DDS_KIND_WRITER: {
-      struct dds_writer const *const wr = (struct dds_writer *)e;
-      // only if SHM is enabled correctly (i.e. iox publisher is initialized) and the type is fixed
-      ret = (wr->m_iox_pub != NULL) && (wr->m_topic->m_stype->fixed_size);
-      break;
-    }
-    default:
-      break;
-  }
-
-  dds_entity_unpin(e);
-#endif
-  (void) entity;
-  return ret;
-}
-
-bool is_loan_available(const dds_entity_t entity)
-{
-  return dds_is_loan_available(entity);
 }
